// src/api-client/api.ts
export interface UserStatsDto {
    'userId': string;
    'count': number;
    'limit': number;
    'remaining': number;
    'resetAt': string;
    'lastRequest': string | null;
}
// ...
export class UsersApi extends BaseAPI {
    /**
     * 
     * @param {string} userId 
     * @param {UsersControllerHistoryRangeEnum} [range] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public usersControllerHistory(userId: string, range?: UsersControllerHistoryRangeEnum, options?: RawAxiosRequestConfig) {
        return UsersApiFp(this.configuration).usersControllerHistory(userId, range, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public usersControllerList(options?: RawAxiosRequestConfig) {
        return UsersApiFp(this.configuration).usersControllerList(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} userId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public usersControllerStats(userId: string, options?: RawAxiosRequestConfig) {
        return UsersApiFp(this.configuration).usersControllerStats(userId, options).then((request) => request(this.axios, this.basePath));
    }
}

export const UsersControllerHistoryRangeEnum = {
    Hour: 'hour',
    Day: 'day'
} as const;
export type UsersControllerHistoryRangeEnum = typeof UsersControllerHistoryRangeEnum[keyof typeof UsersControllerHistoryRangeEnum];